Codechange: split currency.h to _type.h and _func.h
