+++ src/cheat_gui.cpp
@@ -12,7 +12,7 @@
 #include "cheat_type.h"
 #include "company_base.h"
 #include "company_func.h"
-#include "currency.h"
+#include "currency_func.h"
 #include "saveload/saveload.h"
 #include "vehicle_base.h"
 #include "textbuf_gui.h"
